refactor(api): decompose routes.go god-file into domain handler files
routes.go drops from 1848 to 151 lines: now just the cache inventory doc,
pagination helper, /health, and a pure registration list. Closure bodies
moved verbatim into named handlers:

- tx_routes.go, block_routes.go, address_routes.go, token_routes.go,
  validator_routes.go, gas_routes.go, market_routes.go
- validate.go: the three coexisting address/hex validators in one place
  with a comment on why their grammars differ
- helpers.go: requireAddressParam/requireTxHashParam (10x/4x duplicated 400
  guards), parseStandardFilter (3x), contractInfoPayload (2x),
  txActivityCounts (2x block-activity rollup), respondInternal (29 sites),
  parseHexBlockNumber

Also:
- cache janitor no longer starts from init(); StartJanitor returns a stop
  func, started in UserRoute (importing the package no longer leaks a
  goroutine into tests)
- delete dead fetchReceiptLogs (zero callers)
- new hermetic tests: cache TTL/singleflight/janitor-stop, rate-limiter
  refill/eviction, byte-exact 400 guard bodies

Route parity verified: registration inventory identical before/after
(43/43). Response shapes, cache keys, TTLs, and error strings unchanged.
